Extract quick search components into separate component
